🧼 Splits motions state to own files
