+++ esp32/lib/ESP32-sveltekit/Gait/idle_state.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <gait/state.h>
+
+class IdleState : public GaitState {
+  protected:
+    const char *name() const override { return "Idle"; }
+};
